# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ErrorBadRequest defines model for ErrorBadRequest.
type ErrorBadRequest struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

// ErrorInternal defines model for ErrorInternal.
type ErrorInternal struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

// WebhookAckResponse defines model for WebhookAckResponse.
type WebhookAckResponse struct {
	Ok bool `json:"ok"`
}

// WebhookPaymentRequest defines model for WebhookPaymentRequest.
type WebhookPaymentRequest struct {
	Amount     string    `json:"amount"`
	Currency   string    `json:"currency"`
	EventId    string    `json:"event_id"`
	OccurredAt time.Time `json:"occurred_at"`
	Type       string    `json:"type"`
}

// WebhookPaymentParams defines parameters for WebhookPayment.
type WebhookPaymentParams struct {
	XWebhookSignature *string `json:"X-Webhook-Signature,omitempty"`
}

// WebhookPaymentJSONRequestBody defines body for WebhookPayment for application/json ContentType.
type WebhookPaymentJSONRequestBody = WebhookPaymentRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Payment webhook
	// (POST /webhooks/payments)
	WebhookPayment(ctx echo.Context, params WebhookPaymentParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// WebhookPayment converts echo context to params.
func (w *ServerInterfaceWrapper) WebhookPayment(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params WebhookPaymentParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "X-Webhook-Signature" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Webhook-Signature")]; found {
		var XWebhookSignature string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-Webhook-Signature, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Webhook-Signature", valueList[0], &XWebhookSignature, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-Webhook-Signature: %s", err))
		}

		params.XWebhookSignature = &XWebhookSignature
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.WebhookPayment(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/webhooks/payments", wrapper.WebhookPayment)

}

type WebhookPaymentRequestObject struct {
	Params WebhookPaymentParams
	Body   *WebhookPaymentJSONRequestBody
}

type WebhookPaymentResponseObject interface {
	VisitWebhookPaymentResponse(w http.ResponseWriter) error
}

type WebhookPayment200JSONResponse WebhookAckResponse

func (response WebhookPayment200JSONResponse) VisitWebhookPaymentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type WebhookPayment400JSONResponse ErrorBadRequest

func (response WebhookPayment400JSONResponse) VisitWebhookPaymentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type WebhookPayment500JSONResponse ErrorInternal

func (response WebhookPayment500JSONResponse) VisitWebhookPaymentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Payment webhook
	// (POST /webhooks/payments)
	WebhookPayment(ctx context.Context, request WebhookPaymentRequestObject) (WebhookPaymentResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// WebhookPayment operation middleware
func (sh *strictHandler) WebhookPayment(ctx echo.Context, params WebhookPaymentParams) error {
	var request WebhookPaymentRequestObject

	request.Params = params

	var body WebhookPaymentJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.WebhookPayment(ctx.Request().Context(), request.(WebhookPaymentRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "WebhookPayment")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(WebhookPaymentResponseObject); ok {
		return validResponse.VisitWebhookPaymentResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}